Experiment 6
There are flight paths between cities. If there is a flight between city A
and city B then there is an edge between the cities. The cost of the edge can be the time that
flight takes to reach city B from A, or the amount of fuel used for the journey. Represent this
as a graph. The node can be represented by airport name or name of the city. Use adjacency MATRIX
representation of the graph.